requirements.txt
numpy~=1.25.0
matplotlib~=3.7.1
pillow~=9.4.0
seaborn~=0.12.2
pandas~=1.5.3